refactor: pull buildChildProps into constructUrl.js
